+++ app/src/main/java/tv/withaibuild/customiuizer/subs/SelectorResultDelivery.kt
@@ -1,0 +1,11 @@
+package tv.withaibuild.customiuizer.subs
+
+object SelectorResultDelivery {
+    fun canDeliverFromSource(sourceIsAdded: Boolean, targetExists: Boolean): Boolean {
+        return sourceIsAdded && targetExists
+    }
+
+    fun canAcceptAtBackStackTarget(targetExists: Boolean, _targetIsAdded: Boolean): Boolean {
+        return targetExists
+    }
+}